Rename `basic` format to `standard` format
